backend/pyproject.toml
[project]
name = "quranref"
version = "2.0.0"
description = "QuranRef backend API service"
authors = [
    {name = "Kashif Iftikhar", email = "kashif@compulife.com.pk"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "pydantic-settings>=2.8.0",
    "typer>=0.15.1",
    "jinja2>=3.1.4",
    "age-orm>=0.1.0",
    "authlib>=1.3.0",
    "PyJWT>=2.8.0",
    "httpx>=0.27",
    "itsdangerous>=2.1.0",
    "sqlalchemy>=2.0",
    "alembic>=1.14",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.18",
    "bandit>=1.8.0",
    "ipython>=8.31.0",
    "pre-commit>=4.0.1",
]
test = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-env>=1.1.5",
    "httpx>=0.27",
]

[project.scripts]
quranref-cli = "quranref.cli:main"

[tool.uv]
package = true

[tool.uv.sources]
age-orm = { path = "/MyWork/Projects/age-forge/age-orm", editable = true }

[tool.setuptools]
packages = ["quranref"]

[tool.bandit]
exclude_dirs = ["tests"]
# B101 = ignore assert_used
# B404 = allow use of sub process module
# B603 = allow calls to Popen with shell=False
skips = ["B101", "B404", "B603"]

[tool.ruff]
line-length = 100
src = ["quranref"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "--cov=quranref --cov-fail-under=50"
testpaths = [
    "tests",
]

filterwarnings = [
    "ignore::DeprecationWarning",
]

env = [
    "QURANREF_TESTING=true",
    "ENVIRONMENT=testing",
]

[tool.coverage.run]
omit = [
    "quranref/commands/*",
    "quranref/cli.py",
    "quranref/views.py",
    "quranref/template_config.py",
    "quranref/template_tags.py",
    "quranref/log_config.py",
    "quranref/data/*",
    "quranref/lib/*",
]
